completed....

+++ auto_ml/views.py
@@ -1,4 +1,4 @@
-from django.http import JsonResponse, HttpRequest
+from django.http import Http404, JsonResponse, HttpRequest
 from django.shortcuts import render
 from django.core.files.storage import FileSystemStorage
 import pandas as pd
@@ -15,8 +15,18 @@ def home_page(request:HttpRequest):
 
         mediaPath = os.path.join(baseDir, 'media')
         # print(os.listdir(mediaPath))
-        for f in os.listdir(mediaPath):
-            os.remove(os.path.join(mediaPath, f))
+        dirSize:int = __getDirSize(mediaPath)
+        size_limit_mb = 200                                                       # max size in mb 
+        size_limit_bytes:int = size_limit_mb * 1024 * 1024
+
+        # print(f'{size_limit_bytes}  {dirSize}')
+        if size_limit_bytes < dirSize:                                            # delete all files when the size goes above size_limi_mb limit
+            try:
+                for f in os.listdir(mediaPath):
+                    os.remove(os.path.join(mediaPath, f))
+            except Exception as e:
+                print(e)
+
 
         file_json = None        # to convert pd.Dataframe to json
 
@@ -33,7 +43,9 @@ def home_page(request:HttpRequest):
                 return JsonResponse({'error': 'Unsupported file format'}, status=400)
             
             file_json = json.dumps(list(file.columns))      # convert columns to json
-                
+        
+        except FileNotFoundError as e:
+            return render(request, '404.html')
         except Exception as e:
             return JsonResponse({'error': f'Failed to process file: {str(e)}'}, status=400)
         
@@ -43,3 +55,18 @@ def home_page(request:HttpRequest):
         return JsonResponse({'file_id': file_uuid,'data': file_json})    # return json response of the columns of the dataset
 
     return render(request, 'web/index.html')
+
+def __getDirSize(dirname) -> int:
+    size:int = 0
+
+    for root, dirs, files in os.walk(dirname):
+        for file in files:
+            size += os.path.getsize(os.path.join(root,file))
+    
+    return size
+
+def custom_404(request, exception):
+    return render(request, '404.html', status=404)
+
+def custom_500(request):
+    return render(request, '500.html', status=500)
